Controls für Mitglied Liste und Mitglied Detail Views getrennt
