+++ src/main/java/de/jost_net/JVerein/gui/control/AbstractJVereinControl.java
@@ -6,6 +6,7 @@ import de.jost_net.JVerein.gui.dialogs.TabelleSpaltenAuswahlDialog;
 import de.jost_net.JVerein.gui.dialogs.AbstractPartExportDialog.ExportArt;
 import de.jost_net.JVerein.gui.dialogs.TabelleProfilAuswahlDialog;
 import de.jost_net.JVerein.gui.parts.IJVereinPart;
+import de.willuhn.datasource.rmi.ObjectNotFoundException;
 import de.willuhn.jameica.gui.AbstractControl;
 import de.willuhn.jameica.gui.AbstractView;
 import de.willuhn.jameica.gui.GUI;
@@ -33,6 +34,10 @@ public abstract class AbstractJVereinControl extends AbstractControl
       {
         throw e;
       }
+      catch (ObjectNotFoundException e)
+      {
+        throw new ApplicationException("Keine Tabelle vorhanden!");
+      }
       catch (Exception e)
       {
         Logger.error("Fehler beim Profil-Auswahl-Dialog", e);
@@ -51,6 +56,10 @@ public abstract class AbstractJVereinControl extends AbstractControl
       catch (OperationCanceledException | ApplicationException e)
       {
         throw e;
+      }
+      catch (ObjectNotFoundException e)
+      {
+        throw new ApplicationException("Keine Tabelle vorhanden!");
       }
       catch (Exception e)
       {
@@ -85,6 +94,10 @@ public abstract class AbstractJVereinControl extends AbstractControl
           catch (OperationCanceledException | ApplicationException e)
           {
             throw e;
+          }
+          catch (ObjectNotFoundException e)
+          {
+            throw new ApplicationException("Keine Tabelle vorhanden!");
           }
           catch (RemoteException e)
           {
