+++ src/main/java/de/jost_net/JVerein/gui/view/AbstractMitgliedDetailView.java
@@ -417,8 +417,9 @@ public abstract class AbstractMitgliedDetailView extends AbstractDetailView
 
     buttons.paint(parent);
 
-    GUI.getView().addPanelButton(control.exportDetailButton(ExportArt.PDF));
-    GUI.getView().addPanelButton(control.exportDetailButton(ExportArt.CSV));
+    GUI.getView().addPanelButton(control.getProfilePanelButton());
+    GUI.getView().addPanelButton(control.exportButton(ExportArt.PDF));
+    GUI.getView().addPanelButton(control.exportButton(ExportArt.CSV));
     GUI.getView().addPanelButton(control.getSpaltenDetailPanelButton());
   }
