+++ src/main/java/de/jost_net/JVerein/io/AuswertungMitgliedAbstract.java
@@ -14,7 +14,7 @@
 package de.jost_net.JVerein.io;
 
 import de.jost_net.JVerein.Queries.MitgliedQuery.MitgliedAuswahl;
-import de.jost_net.JVerein.gui.control.MitgliedControl;
+import de.jost_net.JVerein.gui.control.MitgliedListeControl;
 import de.jost_net.JVerein.keys.VorlageTyp;
 import de.jost_net.JVerein.util.VorlageUtil;
 
@@ -24,7 +24,7 @@ public abstract class AuswertungMitgliedAbstract implements Exporter
   @Override
   public String getTitle(Object object)
   {
-    if (((MitgliedControl) object).getMitgliedAuswahl()
+    if (((MitgliedListeControl) object).getMitgliedAuswahl()
         .equals(MitgliedAuswahl.MITGLIEDER))
     {
       return VorlageUtil.getName(VorlageTyp.AUSWERTUNG_MITGLIED_TITEL, object,
@@ -40,7 +40,7 @@ public abstract class AuswertungMitgliedAbstract implements Exporter
   @Override
   public String getSubtitle(Object object)
   {
-    if (((MitgliedControl) object).getMitgliedAuswahl()
+    if (((MitgliedListeControl) object).getMitgliedAuswahl()
         .equals(MitgliedAuswahl.MITGLIEDER))
     {
       return VorlageUtil.getName(VorlageTyp.AUSWERTUNG_MITGLIED_SUBTITEL,
